Move core BaseScreen implementation to UITestsFoundation
