+++ WordPress/WordPressUITests/Screens/ActionSheetComponent.swift
@@ -1,4 +1,4 @@
-import Foundation
+import UITestsFoundation
 import XCTest
 
 class ActionSheetComponent: BaseScreen {
