+++ WordPress/WordPressUITests/Screens/ActivityLogScreen.swift
@@ -1,4 +1,4 @@
-import Foundation
+import UITestsFoundation
 import XCTest
 
 class ActivityLogScreen: BaseScreen {
