+++ WordPress/WordPressUITests/Screens/Editor/AztecEditorScreen.swift
@@ -1,4 +1,4 @@
-import Foundation
+import UITestsFoundation
 import XCTest
 
 class AztecEditorScreen: BaseScreen {
