+++ WordPress/WordPressUITests/Screens/Editor/BlockEditorScreen.swift
@@ -1,4 +1,4 @@
-import Foundation
+import UITestsFoundation
 import XCTest
 
 class BlockEditorScreen: BaseScreen {
