+++ WordPress/WordPressUITests/Screens/Editor/EditorNoticeComponent.swift
@@ -1,4 +1,4 @@
-import Foundation
+import UITestsFoundation
 import XCTest
 
 class EditorNoticeComponent: BaseScreen {
