+++ WordPress/WordPressUITests/Screens/Editor/EditorPostSettings.swift
@@ -1,4 +1,4 @@
-import Foundation
+import UITestsFoundation
 import XCTest
 
 class EditorPostSettings: BaseScreen {
