+++ WordPress/WordPressUITests/Screens/Editor/EditorPublishEpilogueScreen.swift
@@ -1,4 +1,4 @@
-import Foundation
+import UITestsFoundation
 import XCTest
 
 class EditorPublishEpilogueScreen: BaseScreen {
