+++ WordPress/WordPressUITests/Screens/Editor/EditorSettingsComponents/CategoriesComponent.swift
@@ -1,4 +1,4 @@
-import Foundation
+import UITestsFoundation
 import XCTest
 
 class CategoriesComponent: BaseScreen {
