+++ WordPress/WordPressUITests/Screens/Editor/EditorSettingsComponents/TagsComponent.swift
@@ -1,4 +1,4 @@
-import Foundation
+import UITestsFoundation
 import XCTest
 
 class TagsComponent: BaseScreen {
