+++ WordPress/WordPressUITests/Screens/Editor/FeaturedImageScreen.swift
@@ -1,4 +1,4 @@
-import Foundation
+import UITestsFoundation
 import XCTest
 
 class FeaturedImageScreen: BaseScreen {
