+++ WordPress/WordPressUITests/Screens/FancyAlertComponent.swift
@@ -1,4 +1,4 @@
-import Foundation
+import UITestsFoundation
 import XCTest
 
 class FancyAlertComponent: BaseScreen {
