+++ WordPress/WordPressUITests/Screens/JetpackBackupOptionsScreen.swift
@@ -1,4 +1,4 @@
-import Foundation
+import UITestsFoundation
 import XCTest
 
 class JetpackBackupOptionsScreen: BaseScreen {
