+++ WordPress/WordPressUITests/Screens/JetpackBackupScreen.swift
@@ -1,4 +1,4 @@
-import Foundation
+import UITestsFoundation
 import XCTest
 
 private struct ElementStringIDs {
