+++ WordPress/WordPressUITests/Screens/JetpackScanScreen.swift
@@ -1,4 +1,4 @@
-import Foundation
+import UITestsFoundation
 import XCTest
 
 class JetpackScanScreen: BaseScreen {
