+++ WordPress/WordPressUITests/Screens/Login/Unified/PrologueScreen.swift
@@ -1,4 +1,4 @@
-import Foundation
+import UITestsFoundation
 import XCTest
 
 private struct ElementStringIDs {
