+++ WordPress/WordPressUITests/Screens/MeTabScreen.swift
@@ -1,4 +1,4 @@
-import Foundation
+import UITestsFoundation
 import XCTest
 
 class MeTabScreen: BaseScreen {
