+++ WordPress/WordPressUITests/Screens/Media/MediaPickerAlbumScreen.swift
@@ -1,3 +1,4 @@
+import UITestsFoundation
 import XCTest
 
 class MediaPickerAlbumScreen: BaseScreen {
