+++ WordPress/WordPressUITests/Screens/MediaScreen.swift
@@ -1,4 +1,4 @@
-import Foundation
+import UITestsFoundation
 import XCTest
 
 class MediaScreen: BaseScreen {
