+++ WordPress/WordPressUITests/Screens/MySitesScreen.swift
@@ -1,4 +1,4 @@
-import Foundation
+import UITestsFoundation
 import XCTest
 
 class MySitesScreen: BaseScreen {
