+++ WordPress/WordPressUITests/Screens/Signup/SignupCheckMagicLinkScreen.swift
@@ -1,4 +1,4 @@
-import Foundation
+import UITestsFoundation
 import XCTest
 
 private struct ElementStringIDs {
