+++ WordPress/WordPressUITests/Screens/Signup/SignupEpilogueScreen.swift
@@ -1,4 +1,4 @@
-import Foundation
+import UITestsFoundation
 import XCTest
 
 private struct ElementStringIDs {
