+++ WordPress/WordPressUITests/Screens/SiteSettingsScreen.swift
@@ -1,4 +1,4 @@
-import Foundation
+import UITestsFoundation
 import XCTest
 
 class SiteSettingsScreen: BaseScreen {
