+++ WordPress/WordPressUITests/Screens/TabNavComponent.swift
@@ -1,4 +1,4 @@
-import Foundation
+import UITestsFoundation
 import XCTest
 
 class TabNavComponent: BaseScreen {
